4.20测试结果
1.用遥控器控制底盘的双odrive的motor0成功.
2.能够读到2006的电机数据了.
3.还没和袁野做仔细的通信协议.

+++ New_DR_Base/Inc/odrive_can.h
@@ -18,7 +18,7 @@ volatile	uint32_t Current_State;
 	
 	float Encoder_Pos_Estimate;
 	float Encoder_Vel_Estimate;
-	int32_t Input_Vel;
+	float Input_Vel;
 	float Vbus_Voltage;
 	
 	uint8_t CMD;
